+++ src/Aspire.Dashboard/Components/Pages/Resources.razor.cs
@@ -310,7 +310,7 @@ public partial class Resources : ComponentBase, IAsyncDisposable, IPageWithSessi
         }
 
         var activeResources = _resourceByName.Values.Where(Filter).OrderBy(e => e.ResourceType).ThenBy(e => e.Name).ToList();
-        var resources = activeResources.Select(r => ResourceDto.MapDto(r, _resourceByName, ColumnsLoc)).ToList();
+        var resources = activeResources.Select(r => ResourceGraphMapper.MapResource(r, _resourceByName, ColumnsLoc)).ToList();
         await _jsModule.InvokeVoidAsync("updateResourcesGraph", resources);
     }
